website_code/requirements.txt
requests==2.27.1
tensorflow==2.8.0
streamlit==1.7.0
google_api_python_client==2.39.0
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2